Predicted melee sounds and effects (#12569)
I didn't want to do this without lag comp. Need to see how it goes in a live environment with moving targets.
